+++ tests/browser-automated-tests-playwright/tests/264_subscriber_order_plan_with_choice_options_and_required_note.js
