+++ tests/browser-automated-tests-playwright/tests/293_subscriber_order_plan_with_recurring_and_upfront_charge.js
@@ -102,6 +102,6 @@ test.describe("order plan with recurring and upfront charge test:", () => {
         const content_upcoming_invoice_plan_upfront_amount = await page.textContent('.plan-sell-price');
         expect(content_upcoming_invoice_plan_upfront_amount === '£1.00')
     });
-    transaction_filter_by_name_and_by_plan_title = require('./transaction_filter_by_name_and_by_plan_title.js');
-    pause_resume_and_cancel_subscriptions = require('./pause_resume_and_cancel_subscriptions.js');
+    transaction_filter_by_name_and_by_plan_title = require('./shop_owner_transaction_filter_by_name_and_by_plan_title.js');
+    pause_resume_and_cancel_subscriptions = require('./shop_owner_pause_resume_and_cancel_subscriptions.js');
 });
