+++ tests/browser-automated-tests-playwright/tests/387_shop_owner_change_shop_colour.js
