+++ tests/browser-automated-tests-playwright/tests/452_shop_owner_categories_creation.js
